Created RoomUserDao and moved some logic there. Also refactored other parts of the app.
