+++ app/src/main/java/com/clover/studio/exampleapp/data/AppDatabase.kt
@@ -27,6 +27,7 @@ abstract class AppDatabase : RoomDatabase() {
     abstract fun chatRoomDao(): ChatRoomDao
     abstract fun messageRecordsDao(): MessageRecordsDao
     abstract fun notesDao(): NotesDao
+    abstract fun roomUserDao(): RoomUserDao
 
     class TablesInfo {
         companion object {
